refactor: Move Globals.ts to the root of src
