+++ packages/tools/devtools/devtools-browser-extension/src/Globals.ts
