+++ packages/tools/devtools/devtools-browser-extension/src/background/fence.json
@@ -1,4 +1,4 @@
 {
 	"tags": ["background"],
-	"imports": ["messaging", "utilities"]
+	"imports": ["globals", "messaging"]
 }
