+++ packages/tools/devtools/devtools-browser-extension/src/content/fence.json
@@ -1,4 +1,4 @@
 {
 	"tags": ["content"],
-	"imports": ["messaging", "utilities"]
+	"imports": ["globals", "messaging"]
 }
