+++ packages/tools/devtools/devtools-browser-extension/src/fence.json
@@ -1,0 +1,4 @@
+{
+	"tags": ["globals"],
+	"exports": ["Globals"]
+}
