+++ packages/tools/devtools/devtools-browser-extension/src/messaging/fence.json
@@ -1,5 +1,5 @@
 {
 	"tags": ["messaging"],
 	"exports": ["index"],
-	"imports": ["utilities"]
+	"imports": []
 }
