+++ packages/tools/devtools/devtools-browser-extension/src/popup/fence.json
@@ -1,4 +1,4 @@
 {
 	"tags": ["popup"],
-	"imports": ["utilities"]
+	"imports": ["globals"]
 }
