--- packages/tools/devtools/devtools-browser-extension/src/utilities/fence.json
@@ -1,5 +1,0 @@
-{
-	"tags": ["utilities"],
-	"exports": ["index", "Globals"],
-	"imports": []
-}
