--- packages/tools/devtools/devtools-browser-extension/src/utilities/index.ts
@@ -1,9 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-// This directory exists to surface utilities, so we want to export everything.
-/* eslint-disable no-restricted-syntax */
-
-export * from "./Globals";
